Task #3 : go to cart & remove from cart
